requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Playwright==1.40.0
requests==2.31.0
python-dotenv==1.0.0
waitress>=3.0.0
psutil>=5.9.0
